create model package and move data into it
